requirements.txt
praw
torch
python-dotenv
python-dotenv
google-generativeai
transformers